build(deps): update to `kube` 0.78 and `k8s-openapi` 0.17 (#42)
This branch updates the dependency on `kube` to v0.78 and the dependency
on `k8s-openapi` to 0.17. Since these dependencies have to be updated
together, the Dependabot PRs for the update don't compile.

Therefore, this PR closes #37 and #36.

It was also necessary to update the devcontainer to the latest version,
as `dev:v34` can no longer build the latest `openssl-sys` (a dependency
of `kube-runtime`). The CI `integration` workflow was also updated to
test against the latest Kubernetes version (v1.26).