Add BackendObjectReference as an inner type
BackendRef is a shared type used by routes that may match traffic to an
arbitrary backend (other than the `Service` they are attached to). Our
current bindings take a simplified approach of representing this by
assuming the type is always a `Service`. Several fields are missing
(gvk, namespace). This makes it hard to work with the types in code,
and it makes it hard to add validation logic since the unserialized
objects will be missing the gvk reference.

This change removes the fields in favour of adding an inner
`BackendObjectReference` type that already exists in the bindings. This
will make it so that any serialization takes into account the actual
type of the backend the route has.

Signed-off-by: Matei David <matei@buoyant.io>