update to `kube-rs` v0.85 and `k8s-openapi` v0.19
This branch updates the dependency on `k8s-openapi` to v0.19 and the
dependency on `kube-rs` to v0.85. This release of `k8s-openapi` also
adds support for Kubernetes v1.27, so I've updated the CI integration
workflow to run tests against that Kubernetes version as well as the
existing ones.

Since this is a breaking change (these are public API deps), I've gone
ahead and bumped the crate version to v0.13.0, as well.